+++ NEWS.md
@@ -3,7 +3,9 @@
 * Added `wrcc_PM2.5_sites_1000`
 * Updated `wrcc_PM2.5_sites_1000` for archival locations.
 * Updated to latest operational `airnow_PM2.5_sites`.
-* Added missing address and elevations to `airnow_PM2.5_sites`.
+* Added missing addresses and elevations to `airnow_PM2.5_sites`.
+* Added missing addresses and elevations to `airsis_PM2.5_sites_1000`.
+* Added missing addresses and elevations to `wrcc_PM2.5_sites_1000`.
 
 # known-locations 0.0.3