+++ NEWS.md
@@ -1,3 +1,9 @@
+# known-locations 0.0.5
+
+March 31, 2022 update.
+
+* 
+
 # known-locations 0.0.4
 
 * Added `wrcc_PM2.5_sites_1000`